Add coverage for todo add edit screen
