+++ lib/screens/add_edit/forms/barrel.dart
@@ -1,1 +1,3 @@
 export 'car_autocomplete.dart';
+export 'repeat_autocomplete.dart';
+export 'cars_checkbox.dart';
